Add ZenGround0's changes to if statements
ZenGround0 (and I to a much lesser extent), made a number
of changes to the if guards in the update_state_verifier
circuit after
we learned more about circom at the IC3 Bootcamp. We thank
Nicola Greco and Dankrad Feist for their insights. The
demo from the bootcamp is at https://github.com/therealyingtong/RollupNC.git
under finish/demo-prep.

We note that circom is both a metalanguage for witness generation
*and* compiles into arithmetic circuits.
An example of problematic if-guards whose behavior
depends on the value of an input is outlined in
the last part of https://github.com/therealyingtong/RollupNC_tutorial.
Helper circuits in helpers/if-gadgets have been added.